clientflow/requirements.txt
Django>=2
django-extensions
djangorestframework
channels
channels_redis
